backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
setuptools==80.9.0
numpy==1.26.4
tensorflow==2.16.2
tensorflow-hub==0.16.1
requests==2.34.0
python-dotenv==1.2.2
faster-whisper==1.1.1
